Integrated display info with watcher graph. Display info is dynamic, loaded, and saved. All unit tests pass. Added boost::regex to libs, so ./autogen.sh && ./configure is needed.

+++ src/libwatcher/labelMessage.h
@@ -15,7 +15,7 @@ namespace watcher {
             public:
                 // The data
                 std::string label;
-                int fontSize;
+                float fontSize;
                 Color foreground;
                 Color background;
                 Timestamp  expiration;    // this should be a float
